+++ R/utils.R
@@ -1,4 +1,6 @@
 
+`%||%` <- function(l, r) if (is.null(l)) r else l
+
 is_interactive <- function() interactive()
 
 is_terminal <- function() {